Revised the "block sustain" number to 1/5
Previously, the requiste number of local representatives to sustain a block of a bill by a party representative was "more than 1/3." This, of course, meant that any bill which could be blocked were necessarily bills that would be legislatively defeated. The purpose of blocks is to allow the minority to enforce a limited amount of power in the legislative process. Requiring more than 1/3 of local representatives to sustain a block would cripple, in most cases, this power of the minority. Thus, I have changed the number to 1/5, which still requires a broad consensus among the minority that the bill should be blocked, but isn't an impossible number to achieve.

In the foregoing, when I say that blocked bills would necessarily be bills that could be legislatively defeated, that was not entirely accurate. As each party representative is alotted a number of votes in legislative votes, there exists a slim margin where more than 1/3 of local representatives would not defeat a bill, however this margin is far too slim to be effective.

For the foregoing reasons, I, Daniel J. Michael Write, modify this drafted constitution.