Provided for life terms of auditors
Previously, the term of auditors was unfixed (thus presumed to be replaceable at any time by their appointors). Further, the court of auditors lacked any provision blocking the ability of auditors to be placed in other offices. To prevent the politicalization, and more importantly, the corruption of the court of auditors, clauses were added to those effects.

-- Daniel J. Michael Write